Added users to entries

+++ server/controllers/entries.js
@@ -4,30 +4,49 @@ const router = require("express").Router();
 const Entry = require("../models/entry");
 
 router.get("/", async (request, response) => {
-  const entries = await Entry.find({});
+  const entries = await Entry.find({}).populate('user', {username: 1, name: 1, admin: 1});
   response.json(entries);
 });
 
 router.post("/", async (request, response) => {
   const entry = new Entry(request.body);
+  const user = request.user;
+
+  if (!user ) {
+    return response.status(403).json({ error: 'user missing' })
+  }
 
   if (!entry.author || !entry.description || !entry.audio) {
     return response.status(400).json({ error: "missing attributes" });
   }
 
+  entry.user = user;
+  user.entries = user.entries.concat(entry._id);
+
+  await user.save();
   const savedEntry = await entry.save();
 
   response.status(201).json(savedEntry);
 });
 
 router.delete("/:id", async (request, response) => {
+  const user = request.user;
+
   console.log(`Removing with id ${request.params.id}`);
   const entry = await Entry.findById(request.params.id);
   if (!entry) {
     return response.status(204).end();
   }
 
+  if ( user.id.toString() !== entry.user.toString() ) {
+    return response.status(403).json({ error: 'user not authorized' })
+  }
+
   await entry.deleteOne();
+
+  user.entries = user.entries.filter(x => x._id.toString() !== entry._id.toString());
+  await user.save()
+
   response.status(204).end();
 });
 
@@ -38,6 +57,7 @@ router.put("/:id", async (request, response) => {
     author: body.author,
     description: body.description,
     audio: body.audio,
+    user: body.user
   };
 
   const updatedEntry = await updatedEntry.findByIdAndUpdate(
